Добавлен тест командной строки.
