+++ hw4/pom.xml
@@ -68,13 +68,6 @@
             <artifactId>spring-boot-configuration-processor</artifactId>
             <optional>true</optional>
         </dependency>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>4.13-beta-3</version>
-            <scope>test</scope>
-        </dependency>
-
     </dependencies>
 
     <build>
