+++ hw4/src/Tests/resources/application.yml
@@ -1,4 +1,7 @@
 spring:
+  shell:
+    interactive:
+      enabled: false
   main:
     banner-mode: console
   messages:
