+++ hw4/src/main/resources/messages.properties
@@ -25,4 +25,5 @@ Shell.Russian=2. Russian
 Shell.You.have.registered=You have registered. Pass the test (test).
 Shell.Show.the.test.result=Show the test result (show).
 Shell.Exit=Exit (exit).
-Shell.First.register= First register (login).
+Shell.First.register=First register (login).
+Shell.Pass.The.Test=First pass the test (test).
