+++ hw4/src/main/resources/messages_ru_RU.properties
@@ -25,6 +25,6 @@ Shell.Russian=2. \u0420\u0443\u0441\u0441\u043A\u0438\u0439
 Shell.You.have.registered=\u0412\u044B \u0437\u0430\u0440\u0435\u0433\u0438\u0441\u0442\u0440\u0438\u0440\u043E\u0432\u0430\u043B\u0438\u0441\u044C. \u041F\u0440\u043E\u0439\u0442\u0438 \u0442\u0435\u0441\u0442  (\u043A\u043E\u043C\u0430\u043D\u0434\u0430 test ).
 Shell.Show.the.test.result=\u041F\u043E\u0441\u043C\u043E\u0442\u0440\u0435\u0442\u044C \u0440\u0435\u0437\u0443\u043B\u044C\u0442\u0430\u0442 \u0442\u0435\u0441\u0442\u0430  (\u043A\u043E\u043C\u0430\u043D\u0434\u0430 show ).
 Shell.Exit=\u0412\u044B\u0439\u0442\u0438  (\u043A\u043E\u043C\u0430\u043D\u0434\u0430 exit ).
-Shell.Pass.The.Test=\u041F\u0440\u043E\u0439\u0434\u0438\u0442\u0435 \u0442\u0435\u0441\u0442  (\u043A\u043E\u043C\u0430\u043D\u0434\u0430 test )
+Shell.Pass.The.Test=\u041F\u0440\u043E\u0439\u0434\u0438\u0442\u0435 \u0442\u0435\u0441\u0442  (\u043A\u043E\u043C\u0430\u043D\u0434\u0430 test)
 Shell.First.register=\u0421\u043D\u0430\u0447\u0430\u043B\u043E \u0437\u0430\u0440\u0435\u0433\u0438\u0441\u0442\u0440\u0438\u0440\u0443\u0439\u0442\u0435\u0441\u044C  (\u043A\u043E\u043C\u0430\u043D\u0434\u0430 login)
